move top level code to config and import result to index
