+++ config/firebase.js
@@ -1,0 +1,1 @@
+console.log('To be added')
